feat: Support poetry_version input to specify Poetry tool version
Reference: https://linear.app/speakeasy/issue/GEN-877/tech-debt-python-poetry-20

This change adds a `poetry_version` input to each of the workflows to specify the Poetry tool version to install and use. By default, the latest (2.0.0 as of this change) is installed, but this is being added as a compatibility layer for calling GitHub Actions workflows that pin the Speakeasy CLI to an older version, which require an older Poetry version before the major version update such as 1.8.3.